feature: list components renamed
